pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "raw_alchemy"
version = "0.1.0"
description = "A Python tool to convert RAW images to ACES and apply creative looks."
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy",
    "imageio",
    "tifffile",
    "click",
    "matplotlib",
    "rawpy @ git+https://github.com/shenmintao/rawpy.git",
    "colour-science @ git+https://github.com/colour-science/colour.git@develop",
    "scipy",
    "Pillow",
    "pyinstaller",
    "numba"
]

[project.urls]
"Homepage" = "https://github.com/shenmintao/raw-alchemy"
"Bug Tracker" = "https://github.com/shenmintao/raw-alchemy/issues"

[project.scripts]
raw-alchemy = "raw_alchemy.cli:main"
raw-alchemy-gui = "raw_alchemy.gui:launch_gui"
